+++ GasExpenses/Sources/CoreData/RefuelPersistentService.swift
@@ -18,58 +18,25 @@ struct RefuelPersistentService: RefuelServiceProtocol, PersistentServiceProtocol
     func addRefuel(_ refuel: Refuel, carID: Int) async throws {
         guard let car = try context.fetch(PersistentCar.fetchRequest()).first(where: { $0.objectID.hash == Int(carID) }) else { return }
 
-        let persistentModel = PersistentRefuel(context: context)
-
-        persistentModel.title = refuel.title
-        persistentModel.comment = refuel.comment
-        persistentModel.car = car
-        if let imageBase64 = refuel.documentBase64 {
-            persistentModel.image = Data(base64Encoded: imageBase64)
-        }
-        if let longitude = refuel.longitude {
-            persistentModel.longitude = longitude
-        }
-
-        if let latitude = refuel.latitude {
-            persistentModel.latitude = latitude
-        }
-
-        persistentModel.date = refuel.date.dateFromJSON()
-        persistentModel.fuelAmount = refuel.fuelAmount
-        persistentModel.mileage = refuel.mileage
-        persistentModel.costPerUnit = refuel.costPerUnit
+        let persistentModel = PersistentRefuel.map(refuel,
+                                                   context: context,
+                                                   car: car)
 
         CoreDataStack.shared.saveContext()
     }
 
     func editRefuel(_ refuel: Refuel) async throws {
-        let persistentModel = PersistentRefuel(context: context)
-
-        persistentModel.title = refuel.title
-        persistentModel.comment = refuel.comment
-        if let imageBase64 = refuel.documentBase64 {
-            persistentModel.image = Data(base64Encoded: imageBase64)
-        }
-        if let longitude = refuel.longitude {
-            persistentModel.longitude = longitude
-        }
-
-        if let latitude = refuel.latitude {
-            persistentModel.latitude = latitude
-        }
-
-        persistentModel.date = refuel.date.dateFromJSON()
-        persistentModel.fuelAmount = refuel.fuelAmount
-        persistentModel.mileage = refuel.mileage
-        persistentModel.costPerUnit = refuel.costPerUnit
-
         let refuels = try context.fetch(PersistentRefuel.fetchRequest())
 
         guard let oldRefuel = refuels.first(where: { $0.objectID.hash == refuel.id }) else { return }
 
-        persistentModel.car = oldRefuel.car
-        context.delete(oldRefuel)
-        try context.save()
+        if let car = oldRefuel.car {
+            let persistentModel = PersistentRefuel.map(refuel,
+                                                       context: context,
+                                                       car: car)
+            context.delete(oldRefuel)
+            try context.save()
+        }
     }
 
     func deleteRefuel(refuelID: Int) async throws {
